fix(messaging): scope Telegram mention status diagnostics
