+++ src/lib/messaging/channels/telegram/manifest.ts
@@ -47,13 +47,6 @@ export const telegramManifest = {
       statePath: "telegramConfig.requireMention",
       validValues: ["0", "1"],
       defaultValue: "1",
-      diagnostics: {
-        doctor: true,
-        valueLabels: {
-          "0": "all-messages",
-          "1": "mention-only",
-        },
-      },
       prompt: {
         label: "Telegram group mention mode",
         help: "Controls Telegram group-chat behavior only — reply only when @mentioned vs. to all group messages. Direct messages are unaffected by this setting and remain subject to pairing and TELEGRAM_ALLOWED_IDS.",
